feat(settings): combined /settings/templates section (#232)
Slice 6 of the Settings redesign (PRD #226). Combines four template-shaped
pages under a single /settings/templates URL with four tabs:

- Estimates       → /settings/estimate-templates body
- Contracts       → /settings/contract-templates body
- Item Library    → /settings/item-library body
- Photo Report Defaults → /settings/reports body (the page is actually
  defaults for the photo-report PDF — preparer name, footer, default
  template, photos per page — so it belongs next to the other
  document-shape configuration)

Old URLs server-redirect to /settings/templates?tab=<key> so bookmarks
keep working. The [id]/edit subroutes for estimate-templates and
contract-templates are left in place — only the list pages move.

Settings sidebar gets one "Templates" entry (LayoutTemplate icon); the
four old entries are removed.
