+++ src/app/settings/templates/estimates-tab.tsx
@@ -1,0 +1,7 @@
+"use client";
+
+import TemplateListClient from "@/components/templates/template-list-client";
+
+export function EstimatesTab() {
+  return <TemplateListClient />;
+}
