+++ src/app/settings/templates/page.tsx
@@ -1,0 +1,25 @@
+"use client";
+
+import { SettingsTabs } from "@/components/settings/settings-tabs";
+import { EstimatesTab } from "./estimates-tab";
+import { ContractsTab } from "./contracts-tab";
+import { ItemLibraryTab } from "./item-library-tab";
+import { PhotoReportDefaultsTab } from "./photo-report-defaults-tab";
+
+export default function TemplatesSettingsPage() {
+  return (
+    <SettingsTabs
+      defaultTab="estimates"
+      tabs={[
+        { key: "estimates", label: "Estimates", content: <EstimatesTab /> },
+        { key: "contracts", label: "Contracts", content: <ContractsTab /> },
+        { key: "item-library", label: "Item Library", content: <ItemLibraryTab /> },
+        {
+          key: "photo-report-defaults",
+          label: "Photo Report Defaults",
+          content: <PhotoReportDefaultsTab />,
+        },
+      ]}
+    />
+  );
+}
