+++ src/lib/settings-nav.ts
@@ -8,7 +8,6 @@ import {
   Download,
   Send,
   CircleDollarSign,
-  Library,
   LayoutTemplate,
 } from "lucide-react";
 import type { ComponentType } from "react";
@@ -24,23 +23,12 @@ export const settingsNavItems: SettingsNavItem[] = [
   { href: "/settings/company", label: "Company Profile", icon: Building2 },
   { href: "/settings/appearance", label: "Appearance", icon: Palette },
   { href: "/settings/jobs", label: "Jobs", icon: ListChecks },
-  { href: "/settings/item-library", label: "Item Library", icon: Library },
-  {
-    href: "/settings/estimate-templates",
-    label: "Estimate Templates",
-    icon: LayoutTemplate,
-  },
+  { href: "/settings/templates", label: "Templates", icon: LayoutTemplate },
   { href: "/settings/pdf-presets", label: "PDF Presets", icon: FileText },
   { href: "/settings/money", label: "Money", icon: CircleDollarSign },
   { href: "/settings/people", label: "People", icon: Users },
   { href: "/settings/email", label: "Email", icon: Mail },
-  {
-    href: "/settings/contract-templates",
-    label: "Contract Templates",
-    icon: FileText,
-  },
   { href: "/settings/contracts", label: "Contracts", icon: Send },
   { href: "/settings/payments", label: "Outgoing Emails", icon: Mail },
-  { href: "/settings/reports", label: "Reports", icon: FileText },
   { href: "/settings/data", label: "Data", icon: Download },
 ];
